# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Kaizen API
  version: 0.1.0
  description: >
    REST API for Kaizen. Single source of truth for endpoints and schemas.
    Payloads use snake_case. All protected routes require a Bearer JWT.
    MVP scope = auth, items, streaks, sync, rule-based redistribute.
    AI endpoints are Phase 1 (premium) and are tagged accordingly.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://api.kaizen.com
    description: Production (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/ai/redistribute:
    post:
      tags: [AI (Phase 1)]
      summary: AI smart redistribution — 2-3 plan variants (premium, Sonnet)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_date]
              properties:
                target_date: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    maxItems: 3
                    items:
                      type: object
                      properties:
                        label: { type: string, example: Balanced day }
                        reason: { type: string }
                        items:
                          type: array
                          items:
                            type: object
                            properties:
                              id: { type: string, format: uuid }
                              scheduled_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/ai/morning-message:
    post:
      tags: [AI (Phase 1)]
      summary: Tone-aware morning review message (premium, Haiku)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pending_count, tone]
              properties:
                pending_count: { type: integer }
                tone: { $ref: '#/components/schemas/Tone' }
                user_name: { type: string }
      responses:
        '200':
          description: 1-2 sentence message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/ai/food-recognize:
    post:
      tags: [AI (Phase 1)]
      summary: Identify food in a photo (premium, Haiku multimodal; max 3/day)
      description: Claude identifies the dish only; КБЖУ comes from the food DB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_base64, media_type]
              properties:
                image_base64: { type: string }
                media_type: { type: string, enum: [image/jpeg, image/png] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  dish: { type: string }
                  portion_description: { type: string }
                  confidence: { type: number, minimum: 0, maximum: 1 }
                  products:
                    type: array
                    description: >
                      Top food-DB matches for the dish (nutrition per 100 g,
                      from Open Food Facts — never from the model). May be empty.
                    items: { $ref: '#/components/schemas/FoodProduct' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429':
          description: Daily food-recognition limit reached
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/ai/wrapped-summary:
    post:
      tags: [AI (Phase 1)]
      summary: One-paragraph wrapped recap of a period (premium)
      description: >
        AI-05. The client computes all numbers from its local DB (code, never
        the model) and sends them; the model only writes a tone-aware
        paragraph (<60 words). On-demand instead of a Sunday cron+Batch
        (ADR-026).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [period_days, tasks_done, tasks_total, main_done, main_total, water_ml, tone]
              properties:
                period_days: { type: integer, minimum: 1, maximum: 366 }
                tasks_done: { type: integer }
                tasks_total: { type: integer }
                main_done: { type: integer }
                main_total: { type: integer }
                avg_mood: { type: number, nullable: true, minimum: 1, maximum: 5 }
                water_ml: { type: integer }
                top_issue: { type: string, nullable: true }
                tone: { $ref: '#/components/schemas/Tone' }
      responses:
        '200':
          description: One upbeat paragraph
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/ai/menu-build:
    post:
      tags: [AI (Phase 1)]
      summary: Compose a one-day menu from the user's own foods (premium)
      description: >
        AI-07, SPEC C5 "[Собрать ИИ]". The client sends its own candidate foods
        and recipes (names + per-100g numbers from the food DB) plus goals; the
        model only picks items and grams — it never outputs nutrition numbers.
        The client recomputes all totals locally (code, never the model).
        Nothing is saved server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [candidates, calorie_goal, protein_goal_g]
              properties:
                candidates:
                  type: array
                  minItems: 5
                  maxItems: 40
                  items:
                    type: object
                    required: [name, per_100g]
                    properties:
                      name: { type: string }
                      per_100g:
                        type: object
                        properties:
                          calories: { type: number, nullable: true }
                          protein: { type: number, nullable: true }
                          fat: { type: number, nullable: true }
                          carbs: { type: number, nullable: true }
                          sugar: { type: number, nullable: true }
                          fiber: { type: number, nullable: true }
                calorie_goal: { type: number, minimum: 800, maximum: 6000 }
                protein_goal_g: { type: number, minimum: 10, maximum: 400 }
                meals:
                  type: array
                  minItems: 1
                  maxItems: 6
                  items: { type: string }
                  default: [breakfast, lunch, dinner]
                tone: { $ref: '#/components/schemas/Tone' }
                health_profile:
                  type: object
                  description: >
                    Optional free-text health notes used only to bias/filter candidate
                    selection. All sub-fields are optional strings (max 500 chars each).
                    Absent = unchanged behavior (backward compatible).
                    Not medical advice — numbers remain code-computed.
                  properties:
                    allergies:
                      type: string
                      maxLength: 500
                      description: Allergies or intolerances the model must avoid.
                    healing:
                      type: string
                      maxLength: 500
                      description: Free-text note about wound healing speed.
                    deficiencies:
                      type: string
                      maxLength: 500
                      description: Known vitamin or mineral deficiencies.
                food_prefs:
                  type: object
                  description: >
                    Optional food preferences used only to filter/bias candidate
                    selection. All sub-fields are optional. Absent = unchanged
                    behavior (backward compatible). Not medical/nutritional
                    prescription — all КБЖУ numbers remain code-computed (ADR-038).
                  properties:
                    diet:
                      type: string
                      maxLength: 300
                      description: >
                        Dietary pattern: none|vegetarian|vegan|pescatarian|
                        halal|kosher|keto|other. "none" or absent = no restriction.
                    goal:
                      type: string
                      maxLength: 300
                      description: >
                        Weight goal: lose|maintain|gain. Already reflected in
                        calorie_goal; included for contextual food biasing only.
                    dislikes:
                      type: string
                      maxLength: 300
                      description: Free-text list of foods or ingredients to avoid.
                    likes:
                      type: string
                      maxLength: 300
                      description: Free-text list of preferred foods (best-effort).
                    meals_per_day:
                      type: integer
                      minimum: 1
                      maximum: 8
                      description: >
                        Target number of meals per day. Informational — the
                        requested meals array remains the source of truth for
                        meal names and count.
      responses:
        '200':
          description: Proposed menu (names + grams only) and a one-sentence note
          content:
            application/json:
              schema:
                type: object
                properties:
                  meals:
                    type: array
                    items:
                      type: object
                      properties:
                        meal: { type: string }
                        items:
                          type: array
                          items:
                            type: object
                            properties:
                              name: { type: string }
                              grams: { type: number }
                  note: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/ai/diary-insight:
    post:
      tags: [AI (Phase 1)]
      summary: Weekly diary insight from recent day logs (premium, Sonnet)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tone]
              properties:
                tone: { $ref: '#/components/schemas/Tone' }
      responses:
        '200':
          description: 2-3 sentence insight
          content:
            application/json:
              schema:
                type: object
                properties:
                  insight: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/ai/schedule-import:
    post:
      tags: [AI (Phase 1)]
      summary: Parse a timetable photo into schedule items (premium, Haiku multimodal)
      description: >
        Claude reads a photo of a timetable/schedule and returns structured items
        (title + scheduled_at) placed on target_date. The client confirms before
        saving (via POST /items). Premium only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_base64, media_type, target_date]
              properties:
                image_base64: { type: string }
                media_type: { type: string, enum: [image/jpeg, image/png] }
                target_date:
                  type: string
                  format: date
                  description: Day to place the parsed items on
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        title: { type: string }
                        scheduled_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not allowed (e.g. free tier on a premium endpoint)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: Email already exists }
    Tone:
      type: string
      enum: [gentle, harsh]
...
    FoodProduct:
      type: object
      description: A product from Open Food Facts; nutrition is per 100 g (null if unknown).
      properties:
        code: { type: string }
        name: { type: string }
        brand: { type: string, nullable: true }
        per_100g:
          type: object
          properties:
            calories: { type: number, nullable: true }
            protein: { type: number, nullable: true }
            fat: { type: number, nullable: true }
            carbs: { type: number, nullable: true }
            sugar: { type: number, nullable: true }
            fiber: { type: number, nullable: true }
...